PrathihasSplitwiseBackend/src/main/java/com/PrathihasProjects/PrathihasSplitwise/helper/Transaction.java
package com.PrathihasProjects.PrathihasSplitwise.helper;


import java.math.BigDecimal;

public class Transaction {
    private String fromUser;
    private String toUser;
    private BigDecimal amount;

    public Transaction(String fromUser, String toUser, BigDecimal amount) {
        this.fromUser = fromUser;
        this.toUser = toUser;
        this.amount = amount;
    }

    // Getters and setters


    public String getFromUser() {
        return fromUser;
    }

    public void setFromUser(String fromUser) {
        this.fromUser = fromUser;
    }

    public String getToUser() {
        return toUser;
    }

    public void setToUser(String toUser) {
        this.toUser = toUser;
    }

    public BigDecimal getAmount() {
        return amount;
    }

    public void setAmount(BigDecimal amount) {
        this.amount = amount;
    }

    @Override
    public String toString() {
        return "Transaction{" +
                "fromUser='" + fromUser + '\'' +
                ", toUser='" + toUser + '\'' +
                ", amount=" + amount +
                '}';
    }
}
